+++ src/Resources/app/administration/src/module/nets-checkout/extension/sw-order/index.js
@@ -72,7 +72,9 @@ Shopware.Component.override('sw-order-user-card', {
                         //
                         me.isLoading = false;
                     });
-            }
+            } else { 
+				me.isLoading = false;
+			}
         },
 
         canRefund() {