+++ CHANGELOG.md
@@ -6,14 +6,25 @@
 |Author | `Nets eCom`
 |Prefix | `EASY-SW6`
 |Shop Version | `6.4+`
-|Version | `1.3.7`
-|Guide | https://tech.nets.eu/shopmodules
+|Version | `1.3.8`
+|Guide | https://developers.nets.eu/nets-easy/en-EU/docs/nets-easy-for-shopware/nets-easy-for-shopware-shopware-6/
 |Github | https://github.com/Nets-eCom/shopware6-easy-checkout
 
 ## CHANGELOG
 
+### Version 1.3.8 - Released 2022-11-14
+* Plugin version update notification api
+* Add support for order items with no product Id
+* Admin Order details page loading icon issue
+
 ### Version 1.3.7 - Released 2022-09-06
 * Improved A2A Payment Methods compatibility
+
+### Version 1.3.6 - Released 2022-05-13
+* Validation Message Fixed for API Test button
+
+### Version 1.3.5 - Released 2022-05-10
+* Minor fixes
 
 ### Version 1.3.4 - Released 2022-04-29
 * Fixed : Issue for order id function for older 6.4 shopware version